Centralize sandbox tool call data
